// WebJuninFrontEnd/src/app/interface/contactos.interface.ts
export interface Contactos {
    id?: number;
    telefono_personal?: string;
    telefono_domicilio?: string;
    telefono_emergencia?: string;
    correo_personal?: string;
    correo_institucional?: string;
    parentesco_emergencia?: string;
    nombre_persona_emergencia?: string;
    funcionario_id?: number;
  }
  